Image_Super-Resolution_ESRGAN/requirements.txt
ipython==6.3.1
ipython-genutils==0.2.0
matplotlib==2.2.2
numpy==1.19.5
Pillow==7.2.0
tensorflow_hub==0.12.0
